[대장간 방어구 강화 기능 추가, 어드민테스트툴 무기/방어구 선택 로직 추가. 건물 업그레이드 기능 추가]
F1: 체력 감소
F2: 스트레스 증가
F3: 테스트용 캐릭터 선택
F4: 테스트용 캐릭터의 무기/방어구 교체
T: 자원 증가.
좌클릭: 건물 건설 및 건물 기능 사용
우클릭: 건물 업그레이드

+++ Assets/97. Moon/V2(Inventory, Building, Supply)/Scripts/BuildingData/Scripts/BuildingData.cs
@@ -10,7 +10,7 @@ public struct BuildingLevelData
     public int oreCost;
     
     [Header("건물 성능 (예: 인벤토리 슬롯 수, 회복량 등)")]
-    public float effectValue; // 수치 
+    public int effectValue; // 수치 
     public int slotExpansion; // 창고 확장
     public int frontierCapAdd; // 개척 데이터 최대치 증가량
     public int materialCapAdd; // 목재/광물 최대치 증가량
